+++ src/common/workload/k8s.py
@@ -49,21 +49,16 @@ class K8sWorkload(AbstractWorkload, ABC):
             return f.read()
 
     @override
-    def write(self, content: str, path: str, mode: str = "w") -> None:
+    def write(self, content: str | bytes, path: str, mode: str = "w") -> None:
         """Writes content to a workload file.
 
         Args:
-            content: string of content to write
+            content: string or bytes content to write
             path: the full filepath to write to
             mode: the write mode. Usually "w" for write, or "a" for append. Default "w"
         """
-        if mode == "a" and (current := self.read(path)):
+        if isinstance(content, str) and mode == "a" and (current := self.read(path)):
             content = current + "\n" + content
-        self.container.push(path, content, make_dirs=True)
-
-    @override
-    def write_bytes(self, content: bytes, path: str) -> None:
-        """Writes bytes content to a workload file."""
         self.container.push(path, content, make_dirs=True)
 
     @override
